PROD-168 updated routing and navigation for new home screen
