+++ app/application/history/page.tsx
@@ -1,6 +1,7 @@
 import { getTotalClaimableRewards } from "@/app/actions/history";
 import History from "@/app/components/History/History";
 import HistoryHeader from "@/app/components/HistoryHeader/HistoryHeader";
+import ProfileNavigation from "@/app/components/ProfileNavigation/ProfileNavigation";
 import { getAllQuestionsReadyForReveal } from "@/app/queries/history";
 
 export default async function Page() {
@@ -10,12 +11,15 @@ export default async function Page() {
   ]);
 
   return (
-    <div className="flex flex-col gap-4 overflow-hidden">
-      <HistoryHeader
-        totalClaimableRewards={totalClaimableRewards}
-        revealableQuestions={revealableQuestions}
-      />
-      <History />
-    </div>
+    <>
+      <ProfileNavigation />
+      <div className="flex flex-col gap-4 overflow-hidden">
+        <HistoryHeader
+          totalClaimableRewards={totalClaimableRewards}
+          revealableQuestions={revealableQuestions}
+        />
+        <History />
+      </div>
+    </>
   );
 }
