+++ app/application/leaderboard/daily/page.tsx
