+++ app/components/ProfileNavigation/constants.ts
@@ -1,5 +1,5 @@
 export const PROFILE_LINKS = [
-  { label: "Dashboard", href: "/application/profile/dashboard" },
-  { label: "Leaderboard", href: "/application/profile/leaderboard" },
-  { label: "History", href: "/application/profile/history" },
+  { label: "Dashboard", href: "/application" },
+  { label: "Leaderboard", href: "/application/leaderboard" },
+  { label: "History", href: "/application/history" },
 ];
